added teacher's template
