+++ Visual/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Visual"
-             StartupUri="View\Student.Forms\MainWindow.xaml">
+             StartupUri="View\Login.Forms\MainWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
