--- Visual/View/Student.Forms/HistoryPage.xaml
@@ -1,17 +1,0 @@
-﻿<Page x:Class="WpfAppTrying.View.Student.HistoryPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:WpfAppTrying.View.Student"
-      mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="800"
-      Title="HistoryPage" Loaded="Page_Loaded">
-
-    <Grid Background="White">
-        <DataGrid x:Name="dgHistory" AutoGenerateColumns="True">
-            
-        </DataGrid>
-
-    </Grid>
-</Page>
