+++ Visual/View/Student.Forms/TaskPage.xaml
@@ -6,10 +6,18 @@
       xmlns:local="clr-namespace:WpfAppTrying.View.Student"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
-      Title="TaskPage">
+      Title="TaskPage" Loaded="Page_Loaded">
 
     <Grid>
-        <TextBlock Text="TASKs" HorizontalAlignment="Center" VerticalAlignment="Center"/>
-
+        <ListBox x:Name="dgHistory">
+            <ListBox.ItemTemplate>
+                <DataTemplate>
+                    <StackPanel Margin="5">
+                        <TextBlock FontSize="16" Text="{Binding Path=Quiz}" HorizontalAlignment="Center" />
+                        <TextBlock FontSize="16" Text="{Binding Path=Teacher}" HorizontalAlignment="Center" />
+                    </StackPanel>
+                </DataTemplate>
+            </ListBox.ItemTemplate>
+        </ListBox>
     </Grid>
 </Page>
