+++ Visual/View/Student.Forms/TaskPage.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,6 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using WpfAppTrying.View.Student.Model;
 
 namespace WpfAppTrying.View.Student
 {
@@ -20,9 +22,20 @@ namespace WpfAppTrying.View.Student
     /// </summary>
     public partial class TaskPage : Page
     {
+        //private BindingList<HistoryList> _historyData;
+        private ObservableCollection<HistoryList> _historyData;
         public TaskPage()
         {
             InitializeComponent();
+        }
+        private void Page_Loaded(object sender, RoutedEventArgs e)
+        {
+            _historyData = new ObservableCollection<HistoryList>()
+            {
+                new HistoryList() { Quiz = "задача 1", Teacher = "Anton" },
+                new HistoryList() { Quiz = "задача 2", Teacher = "Andrey" },
+            };
+            dgHistory.ItemsSource = _historyData;
         }
     }
 }
