+++ Visual/View/Teacher.Form/QuizPage.xaml
@@ -1,0 +1,32 @@
+﻿<Page x:Class="Visual.View.Teacher.Form.QuizPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:Visual.View.Teacher.Form"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="QuizPage">
+
+    <Grid Background="AliceBlue">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+        </Grid.ColumnDefinitions>
+        <ListBox x:Name="lbTestsList" Grid.Column="0">
+            <ListBox.ItemTemplate>
+                <DataTemplate>
+                    <StackPanel Margin="5" Orientation ="Horizontal">
+                        <TextBlock FontSize="16" Text="{Binding Path=Counter}" HorizontalAlignment="Center" />
+                        <TextBlock FontSize="16" Text="{Binding Path=Quiz}" HorizontalAlignment="Center" />
+                    </StackPanel>
+                </DataTemplate>
+            </ListBox.ItemTemplate>
+        </ListBox >
+        <StackPanel Grid.Column="1">
+            <ComboBox x:Name="cbTags" Margin="2" 
+                      ItemsSource="{Binding TagsCollection}" Text="{Binding Tag}"/>
+            <WrapPanel/>
+        </StackPanel>
+    </Grid>
+</Page>
