+++ Visual/View/Teacher.Form/QuizPage.xaml.cs
@@ -10,16 +10,17 @@ using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace Visual
+namespace Visual.View.Teacher.Form
 {
     /// <summary>
-    /// Interaction logic for Window1.xaml
+    /// Interaction logic for QuizPage.xaml
     /// </summary>
-    public partial class Window1 : Window
+    public partial class QuizPage : Page
     {
-        public Window1()
+        public QuizPage()
         {
             InitializeComponent();
         }
