+++ Visual/View/Teacher.Form/QuizTeacherWindow.xaml.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace Visual.View.Teacher.Form
+{
+    /// <summary>
+    /// Interaction logic for QuizTeacher.xaml
+    /// </summary>
+    public partial class QuizTeacher : Window
+    {
+        public QuizTeacher()
+        {
+            InitializeComponent();
+        }
+    }
+}
