+++ Visual/View/Teacher.Form/StatisticTeacherPage.xaml
@@ -1,0 +1,63 @@
+﻿<Page x:Class="Visual.View.Teacher.Form.StatisticTeacher"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:Visual.View.Teacher.Form"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="StatisticTeacher">
+
+    <Grid Background="AliceBlue">
+        
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+        </Grid.ColumnDefinitions>
+        <TreeView Grid.Row="0">
+
+        </TreeView>
+        <!--<DataGrid x:Name="phonesGrid" AutoGenerateColumns="False" HorizontalGridLinesBrush="DarkGray"
+    RowBackground="LightGray" AlternatingRowBackground="White">
+                        
+    
+            <DataGrid.Items>
+                        
+                <local:Phone Title="iPhone 6S" Company="Apple" Price="54990" />
+                        
+                <local:Phone Title="Lumia 950" Company="Microsoft" Price="39990" />
+                        
+                <local:Phone Title="Nexus 5X" Company="Google" Price="29990" />
+                    
+            </DataGrid.Items>
+                
+            <DataGrid.Columns>
+                        
+                <DataGridTextColumn Header="Модель" Binding="{Binding Path=Title}" Width="90" />
+                        
+                <DataGridHyperlinkColumn Header="Компания" Binding="{Binding Path=Company}" Width="80" />
+                        
+                <DataGridTextColumn Header="Цена" Binding="{Binding Path=Price}" Width="50" />
+                    
+            </DataGrid.Columns>
+                 
+    
+            <DataGrid.RowDetailsTemplate>
+                        
+                <DataTemplate>
+                                
+                    <StackPanel Orientation="Horizontal">
+                                        
+                        <TextBlock Text="{Binding Path=Price}" />
+                                        
+                        <TextBlock Text=" рублей по скидке" />
+                                    
+                    </StackPanel>
+                            
+                </DataTemplate>
+                    
+            </DataGrid.RowDetailsTemplate>
+                 
+        </DataGrid>-->
+    </Grid>
+</Page>
