+++ Visual/Visual.csproj
@@ -16,7 +16,6 @@
     
 
     <ItemGroup>
-      <Folder Include="Model\" />
       <Folder Include="ViewModel\" />
       <Folder Include="View\Students.Form\" />
     </ItemGroup>
